break RGRoupAttachmentPoint circular dep
