+++ packages/ketcher-core/src/application/editor/operations/rgroupAttachmentPoint/RGroupAttachmentPointAdd.ts
@@ -7,7 +7,7 @@ import {
 import { OperationPriority, OperationType } from '../OperationType';
 import BaseOperation from '../BaseOperation';
 
-type Data = {
+export type Data = {
   atomId: number;
   attachmentPointType: RGroupAttachmentPointType;
   attachmentPointId?: number;
@@ -15,8 +15,9 @@ type Data = {
 
 class RGroupAttachmentPointAdd extends BaseOperation {
   readonly data: Data;
+  static InverseConstructor: new (attachmentPointId?: number) => BaseOperation;
 
-  constructor(data: Data) {
+  constructor(data: Data = {} as Data) {
     super(
       OperationType.R_GROUP_ATTACHMENT_POINT_ADD,
       OperationPriority.R_GROUP_ATTACHMENT_POINT_ADD,
@@ -55,57 +56,10 @@ class RGroupAttachmentPointAdd extends BaseOperation {
     if (this.data.attachmentPointId === undefined) {
       throw Error(`Inverted attachmentPointId doesn't exist`);
     }
-    // eslint-disable-next-line @typescript-eslint/no-use-before-define
-    return new RGroupAttachmentPointRemove(this.data.attachmentPointId);
+    const inverted = new RGroupAttachmentPointAdd.InverseConstructor();
+    inverted.data = this.data;
+    return inverted;
   }
 }
 
-type RemoveData = {
-  atomId: number;
-  attachmentPointType;
-  attachmentPointId: number;
-};
-
-const INITIAL_DATA = {
-  atomId: 0,
-  attachmentPointType: 1,
-  attachmentPointId: 0,
-};
-
-class RGroupAttachmentPointRemove extends BaseOperation {
-  readonly data: RemoveData;
-
-  constructor(attachmentPointId: number) {
-    super(
-      OperationType.R_GROUP_ATTACHMENT_POINT_REMOVE,
-      OperationPriority.R_GROUP_ATTACHMENT_POINT_REMOVE,
-    );
-    this.data = {
-      ...INITIAL_DATA,
-      attachmentPointId,
-    };
-  }
-
-  execute(restruct: ReStruct) {
-    const { attachmentPointId } = this.data;
-    const struct = restruct.molecule;
-    const item = struct.rgroupAttachmentPoints.get(attachmentPointId);
-    const reItem = restruct.rgroupAttachmentPoints.get(attachmentPointId);
-    assert(item != null && reItem != null);
-
-    this.data.atomId = item.atomId;
-    this.data.attachmentPointType = item.type;
-
-    restruct.markItemRemoved();
-    restruct.clearVisel(reItem.visel);
-    restruct.rgroupAttachmentPoints.delete(attachmentPointId);
-
-    struct.rgroupAttachmentPoints.delete(attachmentPointId);
-  }
-
-  invert() {
-    return new RGroupAttachmentPointAdd(this.data);
-  }
-}
-
-export { RGroupAttachmentPointAdd, RGroupAttachmentPointRemove };
+export { RGroupAttachmentPointAdd };
